+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "private": true,
   "dependencies": {
-    "@fnndsc/chrisapi": "^1.8.0",
+    "@fnndsc/chrisapi": "^1.9.0",
     "@patternfly/react-charts": "^6.88.7",
     "@patternfly/react-core": "^4.235.7",
     "@patternfly/react-styles": "^4.85.7",
