+++ src/api/common/index.ts
@@ -1,4 +1,6 @@
 import * as React from "react";
+import ChrisAPIClient from "../chrisapiclient";
+import { Pipeline, PipelineList, PluginPiping } from "@fnndsc/chrisapi";
 
 export function useSafeDispatch(dispatch: any) {
   const mounted = React.useRef(false);
@@ -162,3 +164,108 @@ export const getFeedTree = (items: any[]) => {
   }
   return tree;
 };
+
+export const fetchPipelines = async (perPage: number, page: number) => {
+  const offset = perPage * (page - 1);
+  const client = ChrisAPIClient.getClient();
+  const params = {
+    limit: perPage,
+    offset: offset,
+  };
+  const registeredPipelinesList = await client.getPipelines(params);
+  const registeredPipelines = registeredPipelinesList.getItems();
+  return {
+    registeredPipelines,
+    registeredPipelinesList,
+  };
+};
+
+
+export async function fetchResources(pipelineInstance: Pipeline) {
+  const params = {
+    limit: 100,
+    offset: 0,
+  };
+
+  const pipelinePluginsFn = pipelineInstance.getPlugins;
+  const pipelineFn = pipelineInstance.getPluginPipings;
+  const boundPipelinePluginFn = pipelinePluginsFn.bind(pipelineInstance);
+  const boundPipelineFn = pipelineFn.bind(pipelineInstance);
+  const pluginPipings: PluginPiping[] = await fetchResource<PluginPiping>(
+    params,
+    boundPipelineFn
+  );
+  const pipelinePlugins: any[] = await fetchResource(
+    params,
+    boundPipelinePluginFn
+  );
+  const parameters = await pipelineInstance.getDefaultParameters({
+    limit: 1000,
+  });
+
+  return {
+    parameters,
+    pluginPipings,
+    pipelinePlugins,
+  };
+}
+
+export const generatePipelineWithName = async (pipelineName: string) => {
+  const client = ChrisAPIClient.getClient();
+  const pipelineInstanceList: PipelineList = await client.getPipelines({
+    name: pipelineName,
+  });
+  const pipelineInstanceId = pipelineInstanceList.data[0].id;
+  const pipelineInstance: Pipeline = await client.getPipeline(
+    pipelineInstanceId
+  );
+  const resources = await fetchResources(pipelineInstance);
+  return {
+    resources,
+    pipelineInstance,
+  };
+};
+
+export const generatePipelineWithData = async (data: any) => {
+  const client = ChrisAPIClient.getClient();
+  const pipelineInstance: Pipeline = await client.createPipeline(data);
+  const resources = await fetchResources(pipelineInstance);
+  return {
+    resources,
+    pipelineInstance,
+  };
+};
+
+export async function fetchComputeInfo(
+  plugin_id: number,
+  dictionary_id: number
+) {
+  const client = ChrisAPIClient.getClient();
+  const computeEnvs = await client.getComputeResources({
+    plugin_id: `${plugin_id}`,
+  });
+
+  if (computeEnvs.getItems()) {
+    const computeEnvData = {
+      [dictionary_id]: {
+        computeEnvs: computeEnvs.data,
+        currentlySelected: computeEnvs.data[0].name,
+      },
+    };
+    return computeEnvData;
+  }
+  return undefined;
+}
+
+export function hasCode(str: string) {
+  let hash = 0;
+  for (let i = 0; i < str.length; i++) {
+    hash = str.charCodeAt(i) + ((hash << 5) - hash);
+  }
+  return hash;
+}
+
+export function intToRGB(i: number) {
+  const c = (i & 0x00ffffff).toString(16).toUpperCase();
+  return "00000".substring(0, 6 - c.length) + c;
+}
