+++ src/components/feed/CreateFeed/LocalFileUpload.tsx
@@ -1,11 +1,11 @@
 import React, { useContext } from 'react'
 import { CreateFeedContext } from './context'
-import { LocalFile, Types } from './types'
+import { LocalFile, Types } from './types/feed'
 import FileUpload from '../../common/fileupload'
 
 const LocalFileUpload: React.FC = () => {
   const { state, dispatch } = useContext(CreateFeedContext)
-  const { localFiles } = state.data
+  const { localFiles } = state.feedState.data
 
   const handleDispatch = (files: LocalFile[]) => {
     dispatch({
