+++ src/components/feed/CreateFeed/utils/pipelines.ts
@@ -32,20 +32,7 @@ export async function fetchResources(pipelineInstance: Pipeline) {
   };
 }
 
-export const fetchPipelines = async (perPage: number, page: number) => {
-  const offset = perPage * (page - 1);
-  const client = ChrisAPIClient.getClient();
-  const params = {
-    limit: perPage,
-    offset: offset,
-  };
-  const registeredPipelinesList = await client.getPipelines(params);
-  const registeredPipelines = registeredPipelinesList.getItems();
-  return {
-    registeredPipelines,
-    registeredPipelinesList,
-  };
-};
+
 
 export const generatePipelineWithName = async (pipelineName: string) => {
   const client = ChrisAPIClient.getClient();
@@ -92,4 +79,17 @@ export async function fetchComputeInfo(
     return computeEnvData;
   }
   return undefined;
+}
+
+export function hasCode(str: string) {
+  let hash = 0;
+  for (let i = 0; i < str.length; i++) {
+    hash = str.charCodeAt(i) + ((hash << 5) - hash);
+  }
+  return hash;
+}
+
+export function intToRGB(i: number) {
+  const c = (i & 0x00ffffff).toString(16).toUpperCase();
+  return "00000".substring(0, 6 - c.length) + c;
 }
