+++ src/pages/DataLibrary/components/UserLibrary/index.tsx
@@ -23,7 +23,7 @@ import LocalSearch from './LocalSearch'
 import { FaUpload } from 'react-icons/fa'
 import FileUpload from '../../../../components/common/fileupload'
 import ChrisAPIClient from '../../../../api/chrisapiclient'
-import { LocalFile } from '../../../../components/feed/CreateFeed/types'
+import { LocalFile } from '../../../../components/feed/CreateFeed/types/feed'
 import { useTypedSelector } from '../../../../store/hooks'
 import { FileSelect, LibraryContext, Types } from './context'
 import { MainRouterContext } from '../../../../routes'
